added version column in LSCONTACT table. Before implementing dynamic columns

+++ app/src/main/java/com/example/muzafarimran/lastingsales/chatheadbubble/BubblesManager.java
@@ -102,9 +102,9 @@ public class BubblesManager {
     }
 
     public void removeBubble(BubbleLayout bubble) {
-//        if (bounded) {
+        if (bounded) {
             bubblesService.removeBubble(bubble);
-//        }
+        }
         mpending = null;
     }
